requirements.txt
django<2.0
djangorestframework<3.7
django-rest-swagger
drf-extensions
psycopg2-binary
whitenoise
django-import-export
django-cors-headers
gunicorn
django-request-logging